Refactor MBeanClient
