+++ app/src/main/java/org/astraea/app/metrics/BeanObject.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.app.metrics.jmx;
+package org.astraea.app.metrics;
 
 import static java.util.Map.Entry;
