+++ app/src/main/java/org/astraea/app/metrics/BeanQuery.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.app.metrics.jmx;
+package org.astraea.app.metrics;
 
 import java.util.HashMap;
 import java.util.Hashtable;
