+++ app/src/main/java/org/astraea/app/metrics/HasBeanObject.java
@@ -16,8 +16,6 @@
  */
 package org.astraea.app.metrics;
 
-import org.astraea.app.metrics.jmx.BeanObject;
-
 public interface HasBeanObject {
   BeanObject beanObject();
