+++ app/src/main/java/org/astraea/app/metrics/broker/ServerMetrics.java
@@ -19,9 +19,9 @@ package org.astraea.app.metrics.broker;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.stream.Collectors;
-import org.astraea.app.metrics.jmx.BeanObject;
-import org.astraea.app.metrics.jmx.BeanQuery;
-import org.astraea.app.metrics.jmx.MBeanClient;
+import org.astraea.app.metrics.BeanObject;
+import org.astraea.app.metrics.BeanQuery;
+import org.astraea.app.metrics.MBeanClient;
 
 public final class ServerMetrics {
