+++ app/src/main/java/org/astraea/app/partitioner/smooth/SmoothWeightRoundRobinDispatcher.java
@@ -38,9 +38,9 @@ import org.astraea.app.common.Utils;
 import org.astraea.app.cost.NeutralIntegratedCost;
 import org.astraea.app.cost.Periodic;
 import org.astraea.app.metrics.HasBeanObject;
+import org.astraea.app.metrics.MBeanClient;
 import org.astraea.app.metrics.collector.BeanCollector;
 import org.astraea.app.metrics.collector.Receiver;
-import org.astraea.app.metrics.jmx.MBeanClient;
 import org.astraea.app.partitioner.Configuration;
 import org.astraea.app.partitioner.Dispatcher;
 import org.astraea.app.partitioner.PartitionerUtils;
