+++ app/src/test/java/org/astraea/app/admin/ClusterBeanTest.java
@@ -18,10 +18,10 @@ package org.astraea.app.admin;
 
 import java.util.List;
 import java.util.Map;
+import org.astraea.app.metrics.BeanObject;
 import org.astraea.app.metrics.KafkaMetrics;
 import org.astraea.app.metrics.broker.HasValue;
 import org.astraea.app.metrics.broker.LogMetrics;
-import org.astraea.app.metrics.jmx.BeanObject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
