+++ app/src/test/java/org/astraea/app/cost/NodeThroughputCostTest.java
@@ -20,8 +20,8 @@ import java.util.List;
 import java.util.Map;
 import org.astraea.app.admin.ClusterBean;
 import org.astraea.app.admin.ClusterInfo;
-import org.astraea.app.metrics.jmx.BeanObject;
-import org.astraea.app.metrics.jmx.MBeanClient;
+import org.astraea.app.metrics.BeanObject;
+import org.astraea.app.metrics.MBeanClient;
 import org.astraea.app.metrics.producer.HasProducerNodeMetrics;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
