+++ app/src/test/java/org/astraea/app/metrics/Utility.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.app.metrics.jmx;
+package org.astraea.app.metrics;
 
 import java.util.Map;
 import javax.management.Attribute;
