+++ app/src/test/java/org/astraea/app/metrics/broker/ServerMetricsTest.java
@@ -16,7 +16,7 @@
  */
 package org.astraea.app.metrics.broker;
 
-import org.astraea.app.metrics.jmx.MBeanClient;
+import org.astraea.app.metrics.MBeanClient;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
