+++ app/src/test/java/org/astraea/app/metrics/collector/FetcherTest.java
@@ -19,7 +19,7 @@ package org.astraea.app.metrics.collector;
 import java.util.List;
 import java.util.Optional;
 import org.astraea.app.metrics.HasBeanObject;
-import org.astraea.app.metrics.jmx.MBeanClient;
+import org.astraea.app.metrics.MBeanClient;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
