+++ app/src/test/java/org/astraea/app/service/RequireJmxServerTest.java
@@ -17,9 +17,9 @@
 package org.astraea.app.service;
 
 import java.util.Set;
+import org.astraea.app.metrics.BeanQuery;
 import org.astraea.app.metrics.KafkaMetrics;
-import org.astraea.app.metrics.jmx.BeanQuery;
-import org.astraea.app.metrics.jmx.MBeanClient;
+import org.astraea.app.metrics.MBeanClient;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
